pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "entari_plugin_hyw"
version = "0.3.0"
description = "Use large language models to interpret chat messages"
authors = [
    {"name" = "kumoSleeping", "email" = "zjr2992@outlook.com"}
]
dependencies = [
    "arclet-entari[full]>=0.16.5",
    "langchain-openai",
    "httpx",
    "bs4>=0.0.2",
    "langchain>=1.0.1",
    "langgraph>=1.0.0",
    "entari-plugin-server>=0.5.0",
    "satori-python-adapter-onebot11>=0.2.5",
    "nekobox>=0.1.6",
    "ddgs>=9.6.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {"text" = "MIT"}
keywords = ["entari", "llm", "ai", "bot", "chat"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/kumoSleeping/entari-plugin-hyw"
Repository = "https://github.com/kumoSleeping/entari-plugin-hyw"
Issues = "https://github.com/kumoSleeping/entari-plugin-hyw/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["entari_plugin_hyw*"]

[tool.setuptools.package-data]
entari_plugin_hyw = ["*.txt", "*.md"]
